Add mapped notification.
